Hooking up apps
linking apps with each other and testing to make sure data is properly set and stored. Home buttons still need to be linked. Service is 100%. Account creation needs to allow 'save and add another' for vehicles and drivers. adjusters and claims work, but need to be linked to the rest of the app and tested.